+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/verification_controller_test.exs
@@ -18,6 +18,7 @@ defmodule BlockScoutWeb.API.V2.VerificationControllerTest do
       assert is_list(response["vyper_compiler_versions"])
       assert is_list(response["verification_options"])
       assert is_list(response["vyper_evm_versions"])
+      assert response["is_rust_verifier_microservice_enabled"] == false
     end
   end